create new config function for authentication mode
